+++ packages/experimental/Avatar/ios/FRNAvatarViewManager.h
@@ -1,0 +1,3 @@
+#import <React/RCTViewManager.h>
+
+#import "FRNAvatarStorage.h"