+++ apps/android/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/tester-android
 
-This log was last generated on Thu, 12 Aug 2021 17:23:44 GMT and should not be manually modified.
+This log was last generated on Fri, 13 Aug 2021 14:54:33 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.1
+
+Fri, 13 Aug 2021 14:54:33 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.35.8 (amdingler694@gmail.com)
+
+## 0.0.1
+
+Fri, 13 Aug 2021 09:05:57 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.35.7 (sanajmi@microsoft.com)
 
 ## 0.0.1