+++ packages/utils/test-tools/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/test-tools
 
-This log was last generated on Thu, 12 Aug 2021 17:23:44 GMT and should not be manually modified.
+This log was last generated on Fri, 13 Aug 2021 16:12:26 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.1
+
+Fri, 13 Aug 2021 16:12:26 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-types to v0.11.1 (t-lindaweng@microsoft.com)
 
 ## 0.1.1